refactor(client): route direct imports through boxElements/layout barrels
Restore the boxElements and layout barrel re-exports that earlier dead-code
cleanup had stripped (they only looked unused because consumers had switched
to direct imports), and route consumers back through the barrels so files that
pulled several siblings from one directory collapse to a single dependency,
lowering their import/max-dependencies count.

- boxElements: re-export CourseMetaItem, DomainTagList, StatusIndicator,
  TopicList; migrate content boxes, tables, and remaining single-symbol
  consumers to the barrel.
- layout: re-export EntityHeaderButton, InfoRow, OverviewCardGrid,
  ResourceLinksSection (with a scoped import/max-dependencies disable, matching
  the api barrel); migrate route consumers to the barrel.
- dashboard dailies body: merge the two direct dailies imports into the
  existing barrel import.

formFields stays direct (barrel<->useAppForm cycle + form.AppField convention);
the deprecated boxes shim and intra-folder imports are left untouched.
