+++ packages/client/src/components/boxElements/index.ts
@@ -1,5 +1,9 @@
 // Cohesive box-display elements grouped so consumers can pull related pieces
 // from one dependency. Add a re-export here when a second consumer needs the
 // same component (a barrel only pays off at >=2 shared sources).
+export { CourseMetaItem } from "./CourseMetaItem";
 export { Description } from "./Description";
+export { DomainTagList } from "./DomainTagList";
 export { EntityLink } from "./EntityLink";
+export { StatusIndicator } from "./StatusIndicator";
+export { TopicList } from "./TopicList";
