+++ packages/client/src/components/contentBoxComponents/ProviderBox.tsx
@@ -2,9 +2,7 @@ import type { CourseProvider } from "@emstack/types";
 
 import { BookIcon, DollarSignIcon, RefreshCwIcon } from "lucide-react";
 
-import { CourseMetaItem } from "@/components/boxElements/CourseMetaItem";
-import { Description } from "@/components/boxElements/Description";
-import { EntityLink } from "@/components/boxElements/EntityLink";
+import { CourseMetaItem, Description, EntityLink } from "@/components/boxElements";
 import {
   ContentBox,
   ContentBoxBody,
