+++ packages/client/src/components/contentBoxComponents/TaskBox.tsx
@@ -3,9 +3,7 @@ import type { Task } from "@emstack/types";
 import { Link } from "@tanstack/react-router";
 import { CheckSquareIcon } from "lucide-react";
 
-import { CourseMetaItem } from "@/components/boxElements/CourseMetaItem";
-import { Description } from "@/components/boxElements/Description";
-import { EntityLink } from "@/components/boxElements/EntityLink";
+import { CourseMetaItem, Description, EntityLink } from "@/components/boxElements";
 import {
   ContentBox,
   ContentBoxBody,
