+++ packages/client/src/components/layout/index.ts
@@ -1,10 +1,15 @@
+/* eslint-disable import/max-dependencies -- barrel re-exporting layout chrome */
 // Barrel for the page-chrome / layout components so a page that renders several
 // of them imports from one module. Layout components import their siblings via
 // direct file paths (never this index), so this re-export introduces no cycle.
 export { DropdownNavItem } from "./DropdownNavItem";
 export { EditForm } from "./EditForm";
 export { EditPageFooter } from "./EditPageFooter";
+export { EntityHeaderButton } from "./EntityHeaderButton";
 export { InfoArea } from "./InfoArea";
+export { InfoRow } from "./InfoRow";
 export { NavDropdown } from "./NavDropdown";
+export { OverviewCardGrid } from "./OverviewCardGrid";
 export { PageHeader } from "./PageHeader";
 export { PageTabs } from "./PageTabs";
+export { ResourceLinksSection } from "./ResourceLinksSection";
