+++ packages/client/src/routes/records.tsx
@@ -2,8 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
 import { BookOpenIcon, Building2Icon, LightbulbIcon } from "lucide-react";
 
-import { OverviewCardGrid } from "@/components/layout/OverviewCardGrid";
-import { PageHeader } from "@/components/layout/PageHeader";
+import { OverviewCardGrid, PageHeader } from "@/components/layout";
 import { ENTITY_DESCRIPTIONS } from "@/lib/entityDescriptions";
 import { fetchProviders, fetchResources, fetchTopics } from "@/utils";
 import { queryKeys } from "@/utils/queryKeys";
